requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
yt-dlp==2024.7.9
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
